Исправлены замечания по ФЗ 9 спринта
