+++ src/adapter/DurationAdapter.java
@@ -1,4 +1,4 @@
-package manager;
+package adapter;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
