+++ src/handler/BaseHttpHandler.java
@@ -1,15 +1,36 @@
-package manager;
+package handler;
 
+import adapter.DurationAdapter;
+import adapter.LocalDateTimeAdapter;
+import com.google.gson.Gson;
+
+import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpExchange;
+import com.sun.net.httpserver.HttpHandler;
+import manager.TaskManager;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
+import java.time.Duration;
+import java.time.LocalDateTime;
 
-public class BaseHttpHandler {
+public abstract class BaseHttpHandler implements HttpHandler {
+
+    protected final TaskManager taskManager;
+    protected final Gson gson;
+
+    protected BaseHttpHandler(TaskManager taskManager, Gson gson) {
+        this.taskManager = taskManager;
+        this.gson = new GsonBuilder()
+                .registerTypeAdapter(LocalDateTime.class, new LocalDateTimeAdapter())
+                .registerTypeAdapter(Duration.class, new DurationAdapter())
+                .create();
+    }
+
     protected void sendText(HttpExchange h, String text, int statusCode) throws IOException {
         byte[] resp = text.getBytes(StandardCharsets.UTF_8);
         h.getResponseHeaders().add("Content-Type", "application/json;charset=utf-8");
-        h.sendResponseHeaders(200, resp.length);
+        h.sendResponseHeaders(statusCode, resp.length);
         h.getResponseBody().write(resp);
         h.close();
     }
@@ -36,6 +57,10 @@ public class BaseHttpHandler {
 
     protected String readRequestBody(HttpExchange exchange) throws IOException {
         return new String(exchange.getRequestBody().readAllBytes(), StandardCharsets.UTF_8);
+    }
+
+    protected void sendBadRequest(HttpExchange exchange, String message) throws IOException {
+        sendText(exchange, "{\"error\": \"" + message + "\"}", 400);
     }
 }
