--- src/manager/PrioritizedHttpHandler.java
@@ -1,35 +1,0 @@
-package manager;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
-import model.Task;
-
-import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.Duration;
-import java.util.List;
-
-public class PrioritizedHttpHandler extends BaseHttpHandler implements HttpHandler {
-    private final TaskManager taskManager;
-    private final Gson gson;
-
-    public PrioritizedHttpHandler(TaskManager taskManager) {
-        this.taskManager = taskManager;
-        this.gson = new GsonBuilder()
-                .registerTypeAdapter(LocalDateTime.class, new LocalDateTimeAdapter())
-                .registerTypeAdapter(Duration.class, new DurationAdapter())
-                .create();
-    }
-
-    @Override
-    public void handle(HttpExchange exchange) throws IOException {
-        if ("GET".equals(exchange.getRequestMethod()) && "/prioritized".equals(exchange.getRequestURI().getPath())) {
-            List<Task> prioritized = taskManager.getPrioritizedTasks();
-            sendSuccess(exchange, gson.toJson(prioritized));
-        } else {
-            sendNotFound(exchange);
-        }
-    }
-}
